Move AxisLeft and AxisBottom to their own files
